Simplify type hierarchy
DefaultFileSystemLocationFingerprint was the only implementation of AbstractFileSystemLocationFingerprint. Now the two types are one.
